name: fix OOB panic on length exceeding data
Upon parsing a malformed DNS packet wherein the name field length
increments the parser's position such that it exceeds the available data
in the packet this will at present result in an out of range indexing
panic.

While it would be best if such malformed panics did not exist, an mDNS
responder can't control what requests it may receive from the network.

Signed-off-by: Grant Elbert <grant.elbert@smartthings.com>