Add with_default_handle_and_ip_list_and_hostname
In some platforms, hostname is fixed so system doesn't gurantee hostname
conflict when the application uses system hostname. This could be the
platform bug but there is a case where the application wants to use specific
and unique hostname for their purpose.

This change add new function to spawn responder task with the application
specific hostname.